TeX Compiler Script Now Captures Important LaTeX Warnings

+++ image/scripts/tex.sh
@@ -45,6 +45,10 @@ rm "$document.synctex.gz" || true
 rm "$document.toc" || true
 rm "$document.vrb" || true
 
+echo "Now removing Unicode BOMs of .tex and .sty files, if any, as they will confuse LaTeX compilers"
+sed -i '1 s/^\xef\xbb\xbf//' "$document.tex"
+sed -i '1 s/^\xef\xbb\xbf//' *.sty || true
+
 echo "We will perform runs of $program/BibTeX until no internal files change anymore."
 
 auxHash=""
@@ -57,7 +61,7 @@ while [ "$oldAuxHash" != "$auxHash" ] || \
       [ "$oldBblHash" != "$bblHash" ]; do
   cycle=$((cycle+1))
   echo "Now beginning build cycle $cycle."
-  
+
   oldAuxHash=$auxHash
   oldBblHash=$bblHash
 
@@ -70,7 +74,7 @@ while [ "$oldAuxHash" != "$auxHash" ] || \
     echo "Error: Program '$program' returned '$retVal' when compiling '$document'. Now exiting."
     exit "$retVal"
   fi
-  
+
   for i in *.aux; do
     if [ "$i" != "$document" ] && \
        [ "$i" != "$document.aux" ] ; then
@@ -85,8 +89,8 @@ while [ "$oldAuxHash" != "$auxHash" ] || \
       fi
     fi
   done
-  
-  
+
+
   if grep -q "\\citation{" "$document.aux"; then
     echo "File '$document.aux' contains citations, so we applying 'bibtex' to it."
     bibtex "$document"
@@ -94,7 +98,7 @@ while [ "$oldAuxHash" != "$auxHash" ] || \
   else
     echo "File '$document.aux' does not contain any citation, so we do not apply 'bibtex'."
   fi
- 
+
   auxHash=""
   for i in *.aux; do
     if [ -f "$i" ]; then
@@ -102,7 +106,7 @@ while [ "$oldAuxHash" != "$auxHash" ] || \
       auxHash="$auxHash$auxHashTemp"
     fi
   done
-  
+
   bblHash=""
   for i in *.bbl; do
     if [ -f "$i" ]; then
@@ -110,9 +114,9 @@ while [ "$oldAuxHash" != "$auxHash" ] || \
       bblHash="$bblHash$bblHashTemp"
     fi
   done
-  
+
   echo "Finished build cycle $cycle."
-    
+
   if (("$cycle" > 200)) ; then
     echo "Something odd is happening: We have performed $cycle cycles. That's too many. Let's quit."
     break
@@ -120,6 +124,39 @@ while [ "$oldAuxHash" != "$auxHash" ] || \
 done
 
 echo "The tool chain '$program'+BibTeX has been executed until nothing changed anymore."
+
+laTeXWarnings=0
+laTeXWarning=""
+if [ -f "$document.log" ]; then
+  echo "We now check that the compilation was successful by 'grep'ing the log file $document.log for common errors/warnings."
+
+  if grep -q "LaTeX Warning: There were undefined references." "$document.log"; then
+    laTeXWarnings=$((laTeXWarnings+1))
+    laTeXWarning="${laTeXWarning}"$'\n'"${laTeXWarnings}. The document contains undefined references. Please fix them (search file $document.log for patterns 'undefined reference' and 'LaTeX Warning: Reference')."
+  fi
+  if grep -q "LaTeX Warning: There were multiply-defined labels." "$document.log"; then
+    laTeXWarnings=$((laTeXWarnings+1))
+    laTeXWarning="${laTeXWarning}"$'\n'"${laTeXWarnings}. The document contains multiply defined labels, i.e., labels defined more than once. Please fix them (search file $document.log for pattern 'multiply-defined')."
+  fi
+  if grep -q "Missing character: There is no" "$document.log"; then
+    laTeXWarnings=$((laTeXWarnings+1))
+    laTeXWarning="${laTeXWarning}"$'\n'"${laTeXWarnings}. The document contains some characters which cannot be printed. Please fix them (check file $document.log for pattern 'Missing character: There is no')."
+  fi
+  if grep -q "Empty ‘thebibliography’ environment" "$document.log"; then
+    laTeXWarnings=$((laTeXWarnings+1))
+    laTeXWarning="${laTeXWarning}"$'\n'"${laTeXWarnings}. The document contains an empty bibliography environment. Maybe you should not use a bibliography if there are no citations? Please fix them (check file $document.log for pattern 'Empty ‘thebibliography’ environment')."
+  fi
+  if grep -q "Float too large for page by" "$document.log"; then
+    laTeXWarnings=$((laTeXWarnings+1))
+    laTeXWarning="${laTeXWarning}"$'\n'"${laTeXWarnings}. At least one floating object such as a table or figure is too large. Please fix them (check file $document.log for pattern 'Float too large for page by')."
+  fi
+  if grep -q "Some font shapes were not available, defaults substituted" "$document.log"; then
+    laTeXWarnings=$((laTeXWarnings+1))
+    laTeXWarning="${laTeXWarning}"$'\n'"${laTeXWarnings}. Some font shapes were unavailable, you should use different characters or fonts. Please fix them (check file $document.log for pattern 'Some font shapes were not available, defaults substituted')."
+  fi
+fi
+
+
 echo "We now ensure that a proper pdf is built."
 
 if [ -f "$document.pdf" ]; then
@@ -131,7 +168,7 @@ else
   if [ ! -f "$document.ps" ]; then
     echo "No postscript (.ps) file '$document.ps' was produced."
     if [ -f "$document.dvi" ]; then
-      echo "The dvi file '$document.dvi' was produced, converting it to postscript (.ps)."      
+      echo "The dvi file '$document.dvi' was produced, converting it to postscript (.ps)."
       dvips "$document"
     fi
   fi
@@ -148,7 +185,9 @@ rm "$document.blg" || true
 rm "$document.dvi" || true
 rm "$document.ent" || true
 rm "$document.idx" || true
-rm "$document.log" || true
+if (("$laTeXWarnings" < 1)) ; then
+  rm "$document.log" || true
+fi
 rm "$document.nav" || true
 rm "$document.out" || true
 rm "$document.out.ps" || true
@@ -161,8 +200,14 @@ rm "$document.toc" || true
 rm "$document.vrb" || true
 
 if [ -f "$document.pdf" ]; then
-  echo "We change the access permissions of the produced document '$document.pdf' to 777." 
+  echo "We change the access permissions of the produced document '$document.pdf' to 777."
   chmod 777 "$document.pdf"
 fi
 
-echo "[La|LuaLa|Pdf|XeLa]TeX compilation finished."
+if (("$laTeXWarnings" < 1)) ; then
+  echo "[La|LuaLa|Pdf|XeLa]TeX compilation finished successfully."
+  exit 0
+else
+  echo "[La|LuaLa|Pdf|XeLa]TeX compilation has not finished successfully, there were some fishy LaTeX warnings. We tried to build a PDF anyway, though:$laTeXWarning"
+  exit 1
+fi
